c:copy_file/mkdir: put in sys.c as they use system calls
